Add new constructor for mc_generic
Added a new constructor for mc_generic which is more C++ like,
namely with a set of parameters. The constructor from a
dictionary is still there with the useless rank removed.
The class is not a template on a dict type anymore, just
the second constructor is. As such, the mc_generic class
should work with no python libs or includes.

  modified:   triqs/mc_tools/mc_generic.hpp
  modified:   triqs/mc_tools/mc_measure_set1.hpp
  modified:   triqs/mc_tools/mc_measure_set2.hpp