gf : cython : remove proxy & wrap block function
- with a simple test in C++ of calling a block function
- removed the proxy.

Decision is as follow :

- all functions returning views  with possible exception
are placed by cython in try catch --> hence the return
type MUST have value semantics.
View a priori  dont.
We can use a proxy, but then all functions in C++ have to return
proxies for no reason.
Besides proxy makes life uneasy in cython (a lot to write, (), <<,
etc).

- Another solution is to allow default constructor of view
and say that if V1,V2 are views :
 if V1 is empty : V1 = V2 just rebinds V1 : V1.rebind(V2)
 else : it copies data into V1.

--> this is not perfect but seems to simplify a lot my life...

--> The defualt constructor for views is private, unles
TRIQS_ALLOW_EMPTY_VIEW is defined (in practice only in pytriqs for
cython module).
So regular C++ code can NEVER make an empty view, which is as dangerous
as a NULL pointer after all...