gf/cython, Demo : Splitting pxd, Demo cython ok
added a file in cmake, which is NOT in the public branch
but which is the cmake directory : FindCythonModuleMacro.cmake

- renamed the module in _gf :
TO DO :  change the construction of modules....
REMOVE THE _!!

- moved many things in the draft/ which are not used.

- Demo : cython now ok. C++ does not compile
(namespace pb, etc...)
Shall I push the solver to triqs/applications ?

- created the pytriqs/pxd directory
for basic cython pxd.
At the moment the gf.pxd is linked there.
A better solution is to add a cython include option in the module ...

---> to do : make a NEW Modulebuild for cython, different from
boost.python....