pade : Fix std::complex compil pb on clang.
- Fix since std::complex can NOT be used
for a generic type : cf :
http://en.cppreference.com/w/cpp/numeric/complex

"
The specializations std::complex<float>, std::complex<double>, and
std::complex<long double> are literal types for representing and
manipulating complex numbers.
The effect of instantiating the template complex for any other type is
unspecified.
"

So we added a new little gmp_complex class.

--> I can not yet remove blitz array because of
a macro ALL defined in blitz conflicting with H5 !!

To do when we get rid of blitz :

- port array (almost done in commented lines)
- change the gf.
- put into a namespace.