Apply new directory structure for pytriqs
Basically the base directory is gone. The dft and solvers have been
moved to applications/dft and applications/impurity_solvers respectively.
All documentation and tests and python scripts have been modified to
take account of these changes. Tests pass. In addition a small
change has been done in FindCythonModuleMacro to allow for the
compilation of the cython module with no installation (useful for
tests).