utility : add apply_on_tuple
If t is a tuple and f a callable object
apply_on_tuple(f,t) is f(*t) in python ...
It unpacks the tuple and call f.
Fully inlined.
Added also speed test, showing same speed as manual code on mac.
To be tested on other compilers before put in usage in TRIQS.
To be use for generic codes.

	new file:   apply_on_tuple_s.cpp
	new file:   ../triqs/utility/CMakeLists.txt
	new file:   ../triqs/utility/apply_on_tuple.cpp
	new file:   ../../triqs/utility/apply_on_tuple.hpp