add pdf documentation generation.
added an option
Build_PDF_Documentation [default OFF]
that build also a pdf documentation of the doc, both for the user and developper manual.

-> sphinx generates latex files, which are then processed by pdflatex.
-> latex files are in doc/developper_manual/latex and doc/user_manual/latex
-> pdf files are doc/developper_manual/latex/TRIQS.pdf and doc/user_manual/latex/TRIQS.pdf
-> pdf files get installed into share/doc/developper_manual/TRIQS.pdf and share/doc/user_manual/TRIQS.pdf

A the moment, the dev. manual compiles ok on my mac, the user manual has one error:
the mott.gif file (gif is not supported by latex).

TO DO :
That is easy to fix, with the "only" directive :
latex -> make a png
html -> keep the gif (because it is animated which only make sense on a web page, I guess ...).

	modified:   CMakeLists.txt
	modified:   cmake/BuildSphinx.cmake
	modified:   doc/developer_manual/CMakeLists.txt
	modified:   doc/user_manual/CMakeLists.txt