gf,py: fix add/sub operation for inhomegenous tail
- when the tail of operand is not of the same size, order_min,
one needs to have _add_ written a bit differently from _iadd_
since a given tail can NOT be resized (it is a view).

- in practice, the add was written as :

def __add__ (self, rhs):
  x = self.copy()
  x += rhs
  return x

but this is NOT always correct :

when self, and rhs have tails of different lengths (e.g. tail starting at 1/omega and a constant !)
then this will fail because of the +=
+= modifies x, but it can not reduce the order_min or the size of a tail (because tail are view in python,
they can be slice, ....)
However, if we write add, we create a new object, here we can guarantee the tail is not a view of something else,
hence we can change directly the _singularity object, and not use the .tail property
which *copies* the rhs into the tail.